+++ __tests__/__integration__/__scripts__/imperative_tso_delete_profile.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -e
-
-echo "================TSO DELETE PROFILE==============="
-zowe profiles delete tso $*
-if [ $? -gt 0 ]
-then
-    exit $?
-fi