+++ packages/imperative/__tests__/__integration__/cmd/__tests__/integration/cli/profiles/Cmd.cli.profile.option.mapping.integration.test.ts
@@ -35,14 +35,14 @@ describe("cmd-cli profile mapping", () => {
         const moldType = "none";
         const response = runCliScript(__dirname + "/__scripts__/profiles/map_banana_to_options.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the profile values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + color);
         expect(response.stdout.toString()).toContain("Description: " + description);
         expect(response.stdout.toString()).toContain("Mold type: " + moldType);
         expect(response.stdout.toString()).toContain("Sweetness: mild");
+        expect(response.status).toBe(0);
     });
 
     it("should have command line arguments take precedence over profile fields", () => {
@@ -55,14 +55,14 @@ describe("cmd-cli profile mapping", () => {
         const cliMoldType = "no mold at all";
         const response = runCliScript(__dirname + "/__scripts__/profiles/banana_profile_and_specify_cli.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType, cliColor, cliDescription, cliMoldType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the CLI arguments
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         expect(response.stdout.toString()).toContain("Sweetness: mild");
+        expect(response.status).toBe(0);
     });
 
     it("should have environmental variables take precedence over profile fields", () => {
@@ -75,14 +75,14 @@ describe("cmd-cli profile mapping", () => {
         const envMoldType = "no mold at all";
         const response = runCliScript(__dirname + "/__scripts__/profiles/banana_profile_and_specify_env.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType, envColor, envDescription, envMoldType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the env variable values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + envColor);
         expect(response.stdout.toString()).toContain("Description: " + envDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + envMoldType);
         expect(response.stdout.toString()).toContain("Sweetness: mild");
+        expect(response.status).toBe(0);
     });
 
     it("should have command line arguments take precedence over profile fields and environmental variables", () => {
@@ -102,14 +102,14 @@ describe("cmd-cli profile mapping", () => {
         const response = runCliScript(__dirname + "/__scripts__/profiles/banana_profile_and_specify_env_and_cli.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType, envColor, envDescription, envMoldType,
                 cliColor, cliDescription, cliMoldType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the CLI arguments
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         expect(response.stdout.toString()).toContain("Sweetness: mild");
+        expect(response.status).toBe(0);
     });
 
     it("should have environmental variables take precedence over default values", () => {
@@ -123,15 +123,15 @@ describe("cmd-cli profile mapping", () => {
 
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_sweetness.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envSweetness]);
-        expect(response.stderr.toString()).toBe("");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the CLI arguments
         // except for sweetness which was specified by environmental variables
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         expect(response.stdout.toString()).toContain("Sweetness: " + envSweetness);
+        expect(response.status).toBe(0);
     });
 
     it("should have service profile fields take precedence over base profile fields", () => {
@@ -140,13 +140,13 @@ describe("cmd-cli profile mapping", () => {
         const kiwiAmount = 1000;
         const response = runCliScript(__dirname + "/__scripts__/profiles/base_and_kiwi_profile.sh",
             TEST_ENVIRONMENT.workingDir, [baseAmount, basePrice, kiwiAmount]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the base profile values
         // except for amount which was specified in service profile
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain(`Amount: ${kiwiAmount}`);
         expect(response.stdout.toString()).toContain(`Price: ${basePrice}`);
+        expect(response.status).toBe(0);
     });
 
     it("should be able to specify positional options via environmental variables", () => {
@@ -156,13 +156,13 @@ describe("cmd-cli profile mapping", () => {
         const envMoldType = "no mold at all";
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_positional.sh",
             TEST_ENVIRONMENT.workingDir, [envColor, envDescription, envMoldType]);
-        expect(response.stderr.toString()).toBe("");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the env variable values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + envColor);
         expect(response.stdout.toString()).toContain("Description: " + envDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + envMoldType);
+        expect(response.status).toBe(0);
     });
 
     it("should map profile fields to positional options", () => {
@@ -171,13 +171,13 @@ describe("cmd-cli profile mapping", () => {
         const moldType = "none";
         const response = runCliScript(__dirname + "/__scripts__/profiles/map_banana_to_positionals.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the profile values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + color);
         expect(response.stdout.toString()).toContain("Description: " + description);
         expect(response.stdout.toString()).toContain("Mold type: " + moldType);
+        expect(response.status).toBe(0);
     });
 
     it("should be able to specify valid number type options via environmental variables", () => {
@@ -188,14 +188,14 @@ describe("cmd-cli profile mapping", () => {
         const envSides = "443";
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_number.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envSides]);
-        expect(response.stderr.toString()).toBe("");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the env variable values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         expect(response.stdout.toString()).toContain("Sides: " + envSides);
+        expect(response.status).toBe(0);
     });
 
     it("should get a syntax error when specifying a non-numeric value via environmental variables", () => {
@@ -207,14 +207,17 @@ describe("cmd-cli profile mapping", () => {
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_number.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envSides]);
 
-        expect(response.stderr.toString()).toContain("failed!");
-        expect(response.stderr.toString()).toContain("Syntax");
-        expect(response.stderr.toString()).toContain("number");
+        expect(response.stderr.toString()).toContain("Syntax Error");
+        expect(response.stderr.toString()).toContain("Invalid value specified for option");
+        expect(response.stderr.toString()).toContain("--sides");
+        expect(response.stderr.toString()).toContain("You specified");
+        expect(response.stderr.toString()).toContain(envSides);
+        expect(response.stderr.toString()).toContain("The value must be a number");
         expect(response.stderr.toString()).toContain(envSides);
         expect(response.status).toBe(1);
     });
 
-    it("should be able to specify valid boolean type options (true) via environmental variables", () => {
+    it("should be able to specify true boolean type options via environmental variables", () => {
         // values used as env variables
         const cliColor = "yellow and black";
         const cliDescription = "A beautiful bunch of ripe banana hides the deadly black tarantula";
@@ -222,17 +225,17 @@ describe("cmd-cli profile mapping", () => {
         const envRipe = "true";
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_boolean.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envRipe]);
-        expect(response.stderr.toString()).toBe("");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the env variable values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         expect(response.stdout.toString()).toContain("Ripe: true");
+        expect(response.status).toBe(0);
     });
 
-    it("should be able to specify valid boolean type options (false) via environmental variables", () => {
+    it("should be able to specify false boolean type options via environmental variables", () => {
         // values used as env variables
         const cliColor = "yellow and black";
         const cliDescription = "A beautiful bunch of ripe banana hides the deadly black tarantula";
@@ -240,14 +243,14 @@ describe("cmd-cli profile mapping", () => {
         const envRipe = "false";
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_boolean.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envRipe]);
-        expect(response.stderr.toString()).toBe("");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the env variable values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         expect(response.stdout.toString()).toContain("Ripe: false");
+        expect(response.status).toBe(0);
     });
 
     it("should get a syntax error when specifying a non-boolean value via environmental variables", () => {
@@ -259,10 +262,12 @@ describe("cmd-cli profile mapping", () => {
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_boolean.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envRipe]);
 
-        expect(response.stderr.toString()).toContain("failed!");
-        expect(response.stderr.toString()).toContain("Syntax");
-        expect(response.stderr.toString()).toContain("boolean");
+        expect(response.stderr.toString()).toContain("Syntax Error");
+        expect(response.stderr.toString()).toContain("Invalid value specified for option");
+        expect(response.stderr.toString()).toContain("--ripe (-r)");
+        expect(response.stderr.toString()).toContain("You specified");
         expect(response.stderr.toString()).toContain(envRipe);
+        expect(response.stderr.toString()).toContain("The value must be a boolean (true or false)");
         expect(response.status).toBe(1);
     });
 
@@ -277,16 +282,16 @@ describe("cmd-cli profile mapping", () => {
         }).join(" ");
         const response = runCliScript(__dirname + "/__scripts__/profiles/specify_env_for_array.sh",
             TEST_ENVIRONMENT.workingDir, [cliColor, cliDescription, cliMoldType, envNames]);
-        expect(response.stderr.toString()).toBe("");
-        expect(response.status).toBe(0);
 
         // the output of the command should use the env variable values
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Color: " + cliColor);
         expect(response.stdout.toString()).toContain("Description: " + cliDescription);
         expect(response.stdout.toString()).toContain("Mold type: " + cliMoldType);
         for (const name of rawNames) {
             expect(response.stdout.toString()).toContain(name.replace("\\'", "'"));
         }
+        expect(response.status).toBe(0);
     });
 
     it("should not map profile fields to --name or --type", () => {
@@ -296,11 +301,12 @@ describe("cmd-cli profile mapping", () => {
         const moldType = "none";
         const response = runCliScript(__dirname + "/__scripts__/profiles/name_type_undefined.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
+
         // name and type should be undefined since we did not specify them via command line
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Name: undefined");
         expect(response.stdout.toString()).toContain("Type: undefined");
+        expect(response.status).toBe(0);
     });
     it("should still be able to specify --name and --type on command line", () => {
         // values used as env variables
@@ -311,10 +317,11 @@ describe("cmd-cli profile mapping", () => {
         const cliType = "Big";
         const response = runCliScript(__dirname + "/__scripts__/profiles/name_type_specify.sh",
             TEST_ENVIRONMENT.workingDir, [color, description, moldType, cliName, cliType]);
-        expect(response.stderr.toString()).toContain("The command 'profiles create' is deprecated.");
-        expect(response.status).toBe(0);
+
         // name and type should be undefined since we did not specify them via command line
+        expect(response.stderr.toString()).toBe("");
         expect(response.stdout.toString()).toContain("Name: " + cliName);
         expect(response.stdout.toString()).toContain("Type: " + cliType);
+        expect(response.status).toBe(0);
     });
 });