+++ packages/imperative/__tests__/__integration__/imperative/__tests__/__integration__/cli/config/convert-profiles/cli.imperative-test-cli.config.convert-profiles.integration.subtest.ts
@@ -12,7 +12,7 @@
 import * as fs from "fs";
 import * as fsExtra from "fs-extra";
 import * as path from "path";
-import { keyring as keytar } from "@zowe/secrets-for-zowe-sdk";
+import { keyring } from "@zowe/secrets-for-zowe-sdk";
 import { ITestEnvironment } from "../../../../../../../__src__/environment/doc/response/ITestEnvironment";
 import { SetupTestEnvironment } from "../../../../../../../__src__/environment/SetupTestEnvironment";
 import { runCliScript } from "../../../../../../../src/TestUtil";
@@ -33,11 +33,14 @@ describe("imperative-test-cli config convert-profiles", () => {
     });
 
     beforeEach(() => {
-        runCliScript(__dirname + "/__scripts__/create_profiles_secured_and_base.sh", TEST_ENVIRONMENT.workingDir);
+        fsExtra.copySync(__dirname + "/../../config/__resources__/profiles_secured_and_base", TEST_ENVIRONMENT.workingDir + "/profiles");
     });
 
     afterEach(() => {
-        runCliScript(__dirname + "/__scripts__/delete_profiles_secured_and_base_noerr.sh", TEST_ENVIRONMENT.workingDir);
+        const response = runCliScript(__dirname + "/__scripts__/delete_profiles.sh", TEST_ENVIRONMENT.workingDir);
+        expect(response.stdout.toString()).toEqual("");
+        expect(response.stderr.toString()).toEqual("");
+
         if (fs.existsSync(configJsonPath)) {
             fs.unlinkSync(configJsonPath);
         }
@@ -53,12 +56,17 @@ describe("imperative-test-cli config convert-profiles", () => {
         });
 
         it("should convert profiles to team config", async () => {
+            // set a value in the secure vault that would have been created for the V1 secured profile
+            await keyring.setPassword("imperative-test-cli", "secured_test_secret",
+                Buffer.from('"world"').toString("base64")
+            );
+
             const response = runCliScript(__dirname + "/__scripts__/convert_profiles.sh", TEST_ENVIRONMENT.workingDir, ["y"]);
-            expect(response.status).toBe(0);
+            expect(response.stderr.toString()).toEqual("");
             expect(response.stdout.toString()).toContain("Detected 2 old profile(s) to convert");
             expect(response.stdout.toString()).toContain("Your new profiles have been saved");
             expect(response.stdout.toString()).toContain("Your old profiles have been moved");
-            expect(response.stderr.toString()).toEqual("");
+            expect(response.status).toBe(0);
 
             // Check contents of config JSON
             const configJson = JSON.parse(fs.readFileSync(configJsonPath, "utf-8"));
@@ -87,7 +95,7 @@ describe("imperative-test-cli config convert-profiles", () => {
             });
 
             // Check secure credentials stored in vault
-            const securedValue = await keytar.getPassword("imperative-test-cli", "secure_config_props");
+            const securedValue = await keyring.getPassword("imperative-test-cli", "secure_config_props");
             const secureConfigProps = JSON.parse(Buffer.from(securedValue, "base64").toString());
             expect(secureConfigProps).toMatchObject({
                 [configJsonPath]: {
@@ -104,10 +112,13 @@ describe("imperative-test-cli config convert-profiles", () => {
 
     it("should convert v1 profile property names to v2 names", async () => {
         // we don't want the profiles created by beforeEach(). We only want an old profile.
-        runCliScript(__dirname + "/__scripts__/delete_profiles_secured_and_base_noerr.sh", TEST_ENVIRONMENT.workingDir);
+        let response = runCliScript(__dirname + "/__scripts__/delete_profiles.sh", TEST_ENVIRONMENT.workingDir);
+        expect(response.stdout.toString()).toEqual("");
+        expect(response.stderr.toString()).toEqual("");
+
         fsExtra.copySync(__dirname + "/../../config/__resources__/profiles_with_v1_names", TEST_ENVIRONMENT.workingDir + "/profiles");
 
-        const response = runCliScript(__dirname + "/__scripts__/convert_profiles.sh", TEST_ENVIRONMENT.workingDir, ["y"]);
+        response = runCliScript(__dirname + "/__scripts__/convert_profiles.sh", TEST_ENVIRONMENT.workingDir, ["y"]);
         expect(response.status).toBe(0);
         expect(response.stdout.toString()).toContain("Detected 1 old profile(s) to convert");
         expect(response.stdout.toString()).toContain("Your new profiles have been saved");
@@ -142,17 +153,20 @@ describe("imperative-test-cli config convert-profiles", () => {
     describe("failure scenarios", () => {
         it("should not convert profiles if prompt is rejected", () => {
             const response = runCliScript(__dirname + "/__scripts__/convert_profiles.sh", TEST_ENVIRONMENT.workingDir, ["n"]);
-            expect(response.status).toBe(0);
+            expect(response.stderr.toString()).toEqual("");
             expect(response.stdout.toString()).toContain("Detected 2 old profile(s) to convert");
             expect(response.stdout.toString()).not.toContain("Your new profiles have been saved");
             expect(response.stdout.toString()).not.toContain("Your old profiles have been moved");
-            expect(response.stderr.toString()).toEqual("");
+            expect(response.status).toBe(0);
             expect(fs.existsSync(configJsonPath)).toBe(false);
         });
         it("should not delete profiles if prompt is rejected", () => {
-            runCliScript(__dirname + "/__scripts__/delete_profiles_secured_and_base.sh", TEST_ENVIRONMENT.workingDir);
+            // delete profiles previously created, but leave the profile type definitions
+            let response = runCliScript(__dirname + "/__scripts__/delete_profiles.sh", TEST_ENVIRONMENT.workingDir);
+            expect(response.stderr.toString()).toEqual("");
+            expect(response.stdout.toString()).toEqual("");
 
-            const response = runCliScript(__dirname + "/__scripts__/convert_profiles_delete.sh", TEST_ENVIRONMENT.workingDir, ["n"]);
+            response = runCliScript(__dirname + "/__scripts__/convert_profiles_delete.sh", TEST_ENVIRONMENT.workingDir, ["n"]);
             expect(response.status).toBe(0);
             expect(response.stdout.toString()).toContain("No old profiles were found");
             expect(response.stdout.toString()).toContain("Are you sure you want to delete your v1 profiles?");